+++ test/models/deploy_spec_test.rb
@@ -67,13 +67,15 @@ module Shipit
       @spec.stubs(:gemfile_lock_exists?).returns(true)
       command = %(
         bundle install
-        --frozen
         --jobs 4
-        --path #{DeploySpec.bundle_path}
         --retry 2
-        --without=default:production:development:test:staging:benchmark:debug
       ).gsub(/\s+/, ' ').strip
+      config_command = "bundle config set --local path #{@spec.bundle_path}"
+      without_command = "bundle config set --local without 'default:production:development:test:staging:benchmark:debug'"
+
       assert_equal command, @spec.bundle_install.last
+      assert @spec.bundle_install.include?(config_command)
+      assert @spec.bundle_install.include?(without_command)
     end
 
     test '#bundle_install use `dependencies.bundler.without` if present to build the --without argument' do
@@ -81,31 +83,30 @@ module Shipit
       @spec.stubs(:load_config).returns('dependencies' => { 'bundler' => { 'without' => %w[some custom groups] } })
       command = %(
         bundle install
-        --frozen
         --jobs 4
-        --path #{DeploySpec.bundle_path}
         --retry 2
-        --without=some:custom:groups
       ).gsub(/\s+/, ' ').strip
       assert_equal command, @spec.bundle_install.last
+      without_command = "bundle config set --local without 'some:custom:groups'"
+      assert @spec.bundle_install.include?(without_command)
     end
 
-    test '#bundle_install has --frozen option if Gemfile.lock is present' do
+    test '#bundle_install configures frozen mode if Gemfile.lock is present' do
       @spec.stubs(:load_config).returns('dependencies' => { 'bundler' => { 'without' => %w[some custom groups] } })
       @spec.stubs(:gemfile_lock_exists?).returns(true)
-      assert @spec.bundle_install.last.include?('--frozen')
+      assert @spec.bundle_install.include?('bundle config set --local frozen true')
     end
 
-    test '#bundle_install does not have --frozen option if Gemfile.lock is not present' do
+    test '#bundle_install does not configure frozen mode if Gemfile.lock is not present' do
       @spec.stubs(:load_config).returns('dependencies' => { 'bundler' => { 'without' => %w[some custom groups] } })
       @spec.stubs(:gemfile_lock_exists?).returns(false)
-      refute @spec.bundle_install.last.include?('--frozen')
+      refute @spec.bundle_install.include?('bundle config set --local frozen true')
     end
 
-    test '#bundle_install does not have --frozen if overridden in shipit.yml' do
+    test '#bundle_install does not configure frozen mode if overridden in shipit.yml' do
       @spec.stubs(:load_config).returns('dependencies' => { 'bundler' => { 'frozen' => false } })
       @spec.stubs(:gemfile_lock_exists?).returns(true)
-      refute @spec.bundle_install.last.include?('--frozen')
+      refute @spec.bundle_install.include?('bundle config set --local frozen true')
     end
 
     test "#provisioning_handler returns `provision.handler` if present" do